git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/smoothvideo@19504 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/UPnP.cpp
@@ -1660,7 +1660,7 @@ CUPnPRenderer::GetMetadata(NPT_String& meta)
                                                        NPT_HttpRequestContext());
     if (object) {
         // fetch the path to the thumbnail
-        CStdString thumb = g_infoManager.GetImage(MUSICPLAYER_COVER, -1); //TODO: Only audio for now
+        CStdString thumb = g_infoManager.GetImage(MUSICPLAYER_COVER, (DWORD)-1); //TODO: Only audio for now
 
 #if defined(HAS_LINUX_NETWORK) || defined(HAS_WIN32_NETWORK)
         NPT_String ip;
